feat: add server selection
Adds server selection to the dashboard page and subviews. Optimizes discord.js usage by storing the client as a singleton on the server object so it can be retrieved in next.js routes and serverProps methods to leverage its existing cache and skip redundant logins. Leverages this plus ActiveGuild context to list servers and track the current selected one.